requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.2.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron_descriptor==1.4.0
Django==4.2.3
django-celery-beat==2.5.0
django-cors-headers==4.1.0
django-debug-toolbar==4.2.0
django-filter==23.2
django-modeltranslation==0.18.10
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.0
drf-yasg==1.21.6
Faker==19.2.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.1
MarkupSafe==2.1.3
msgpack==1.0.5
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.39
`PyJWT==2.7.0
pytest==7.4.0
pytest-django==4.5.2
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.33.0
requests==2.31.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
wheel==0.43.0
